+++ server/src/main/java/com/auction/server/model/AuctionSession.java
@@ -2,6 +2,7 @@ package com.auction.server.model;
 
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Entity
 public class AuctionSession {
@@ -13,9 +14,9 @@ public class AuctionSession {
     @JoinColumn(name = "product_id")
     private Product product;
 
-    private Double startingPrice;
-    private Double currentPrice;
-    private Double stepPrice;
+    private BigDecimal startingPrice;
+    private BigDecimal currentPrice;
+    private BigDecimal stepPrice;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
     private String status;
@@ -26,12 +27,12 @@ public class AuctionSession {
     public void setId(Integer id) { this.id = id; }
     public Product getProduct() { return product; }
     public void setProduct(Product product) { this.product = product; }
-    public Double getStartingPrice() { return startingPrice; }
-    public void setStartingPrice(Double startingPrice) { this.startingPrice = startingPrice; }
-    public Double getCurrentPrice() { return currentPrice; }
-    public void setCurrentPrice(Double currentPrice) { this.currentPrice = currentPrice; }
-    public Double getStepPrice() { return stepPrice; }
-    public void setStepPrice(Double stepPrice) { this.stepPrice = stepPrice; }
+    public BigDecimal getStartingPrice() { return startingPrice; }
+    public void setStartingPrice(BigDecimal startingPrice) { this.startingPrice = startingPrice; }
+    public BigDecimal getCurrentPrice() { return currentPrice; }
+    public void setCurrentPrice(BigDecimal currentPrice) { this.currentPrice = currentPrice; }
+    public BigDecimal getStepPrice() { return stepPrice; }
+    public void setStepPrice(BigDecimal stepPrice) { this.stepPrice = stepPrice; }
     public LocalDateTime getStartTime() { return startTime; }
     public void setStartTime(LocalDateTime startTime) { this.startTime = startTime; }
     public LocalDateTime getEndTime() { return endTime; }
